third_party/blink/web_tests/svg/dom/undefined-null-expected.txt

Check that |undefined| and |null| arguments do not cause crashes, but do throw exceptions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

SVGGeometryElement


isPointInFill(SVGPoint point)
PASS circle.isPointInFill(point) did not throw exception.
PASS circle.isPointInFill(undefined) threw exception TypeError: Failed to execute 'isPointInFill' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'..
PASS circle.isPointInFill(null) threw exception TypeError: Failed to execute 'isPointInFill' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'..


isPointInStroke(SVGPoint point)
PASS circle.isPointInStroke(point) did not throw exception.
PASS circle.isPointInStroke(undefined) threw exception TypeError: Failed to execute 'isPointInStroke' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'..
PASS circle.isPointInStroke(null) threw exception TypeError: Failed to execute 'isPointInStroke' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'..




SVGSVGElement


getIntersectionList(SVGRect rect, SVGElement? referenceElement)
PASS svg.getIntersectionList(rect, svg) did not throw exception.
PASS svg.getIntersectionList(undefined, svg) threw exception TypeError: Failed to execute 'getIntersectionList' on 'SVGSVGElement': parameter 1 is not of type 'SVGRect'..
PASS svg.getIntersectionList(null, svg) threw exception TypeError: Failed to execute 'getIntersectionList' on 'SVGSVGElement': parameter 1 is not of type 'SVGRect'..
PASS svg.getIntersectionList(rect, undefined) did not throw exception.
PASS svg.getIntersectionList(rect, null) did not throw exception.


getEnclosureList(SVGRect rect, SVGElement? referenceElement)
PASS svg.getEnclosureList(rect, svg) did not throw exception.
PASS svg.getEnclosureList(undefined, svg) threw exception TypeError: Failed to execute 'getEnclosureList' on 'SVGSVGElement': parameter 1 is not of type 'SVGRect'..
PASS svg.getEnclosureList(null, svg) threw exception TypeError: Failed to execute 'getEnclosureList' on 'SVGSVGElement': parameter 1 is not of type 'SVGRect'..
PASS svg.getEnclosureList(rect, undefined) did not throw exception.
PASS svg.getEnclosureList(rect, null) did not throw exception.


checkIntersection(SVGElement element, SVGRect rect)
PASS svg.checkIntersection(svg, rect) did not throw exception.
PASS svg.checkIntersection(undefined, rect) threw exception TypeError: Failed to execute 'checkIntersection' on 'SVGSVGElement': parameter 1 is not of type 'SVGElement'..
PASS svg.checkIntersection(null, rect) threw exception TypeError: Failed to execute 'checkIntersection' on 'SVGSVGElement': parameter 1 is not of type 'SVGElement'..
PASS svg.checkIntersection(svg, undefined) threw exception TypeError: Failed to execute 'checkIntersection' on 'SVGSVGElement': parameter 2 is not of type 'SVGRect'..
PASS svg.checkIntersection(svg, null) threw exception TypeError: Failed to execute 'checkIntersection' on 'SVGSVGElement': parameter 2 is not of type 'SVGRect'..


checkEnclosure(SVGElement element, SVGRect rect)
PASS svg.checkEnclosure(svg, rect) did not throw exception.
PASS svg.checkEnclosure(undefined, rect) threw exception TypeError: Failed to execute 'checkEnclosure' on 'SVGSVGElement': parameter 1 is not of type 'SVGElement'..
PASS svg.checkEnclosure(null, rect) threw exception TypeError: Failed to execute 'checkEnclosure' on 'SVGSVGElement': parameter 1 is not of type 'SVGElement'..
PASS svg.checkEnclosure(svg, undefined) threw exception TypeError: Failed to execute 'checkEnclosure' on 'SVGSVGElement': parameter 2 is not of type 'SVGRect'..
PASS svg.checkEnclosure(svg, null) threw exception TypeError: Failed to execute 'checkEnclosure' on 'SVGSVGElement': parameter 2 is not of type 'SVGRect'..


SVGTransform createSVGTransformFromMatrix(SVGMatrix matrix)
PASS svg.createSVGTransformFromMatrix() threw exception TypeError: Failed to execute 'createSVGTransformFromMatrix' on 'SVGSVGElement': 1 argument required, but only 0 present..
PASS svg.createSVGTransformFromMatrix(matrix) did not throw exception.
PASS svg.createSVGTransformFromMatrix(undefined) threw exception TypeError: Failed to execute 'createSVGTransformFromMatrix' on 'SVGSVGElement': parameter 1 is not of type 'SVGMatrix'..
PASS svg.createSVGTransformFromMatrix(null) threw exception TypeError: Failed to execute 'createSVGTransformFromMatrix' on 'SVGSVGElement': parameter 1 is not of type 'SVGMatrix'..




SVGTextContentElement


getCharNumAtPosition(SVGPoint point)
PASS text.getCharNumAtPosition(point) did not throw exception.
PASS text.getCharNumAtPosition(undefined) threw exception TypeError: Failed to execute 'getCharNumAtPosition' on 'SVGTextContentElement': parameter 1 is not of type 'SVGPoint'..
PASS text.getCharNumAtPosition(null) threw exception TypeError: Failed to execute 'getCharNumAtPosition' on 'SVGTextContentElement': parameter 1 is not of type 'SVGPoint'..
PASS successfullyParsed is true

TEST COMPLETE
